requirements.txt
flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
google-generativeai==0.7.2
SpeechRecognition==3.10.4
gTTS==2.5.3
openai-whisper
openai
pydub
torch 
transformers 
accelerate 
bitsandbytes 
soundfile 
langchain 
chromadb 
sentence-transformers
pypdf
